% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastDMinR.R
\name{fast_dm}
\alias{fast_dm}
\title{Run fast-dm}
\usage{
fast_dm(data, Subject, Conditions, TIME, RESPONSE, fast_dm_path, method,
  precision, fix_to, depend_on_condition, invariant, delete = TRUE)
}
\arguments{
\item{data}{Dataframe containing all relevant columns (Subject, Conditions, TIME and RESPONSE).}

\item{Subject}{Name of the column containing the subject id.}

\item{Conditions}{Name(s) of the column(s) containing the levels of conditions across which parameters might vary.}

\item{TIME}{Name of the column containing response times (in seconds).}

\item{RESPONSE}{Name of the column containing responses.}

\item{method}{Method of parameter estimation. Can be "ml", "ks" or "cs" for maximum likelihood, kolmogorov smirnov or chi squared.}

\item{precision}{Precision of calculation. VALUE corresponds roughly to the number of decimals of the predicted CDFs that are calculated accuratly.}

\item{fix_to}{List containing fixed parameter values. Names are parameters, elements are values.}

\item{depend_on_condition}{List containing parameters which may vary across conditions. Names are parameters, elements are conditions.}

\item{invariant}{Vector of names of parameters, which are not fixed but invariant across conditions.}

\item{delete}{Should temporary files be deleted?}
}
\value{
Returns a list of the input data, individual and aggregated parameter estimates as well as individual and aggregated cdf values per condition.
}
\description{
Estimate a diffusion model using fast-dm.
}
\seealso{
For further information regarding fast-dm see \url{https://www.psychologie.uni-heidelberg.de/ae/meth/fast-dm/}.
}
\keyword{diffusion}
\keyword{fast-dm}
\keyword{modeling}
