% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastDMinR.R
\name{fast_dm}
\alias{fast_dm}
\title{Run fast-dm}
\usage{
fast_dm(data, Subject, Conditions, TIME, RESPONSE, fast_dm_path, method,
  precision, fix_to, depend_on_condition, invariant, delete = TRUE)
}
\arguments{
\item{data}{Dataframe containing all relevant columns (Subject, Conditions, TIME and RESPONSE).}

\item{Subject}{Name of the column containing the subject id.}

\item{Conditions}{Name(s) of the column(s) containing the levels of conditions across which parameters might vary.}

\item{TIME}{Name of the column containing response times (in seconds).}

\item{RESPONSE}{Name of the column containing responses.}

\item{method}{Method of parameter estimation. Can be "ml", "ks" or "cs" for maximum likelihood, kolmogorov smirnov or chi squared.}

\item{precision}{Precision of calculation. VALUE corresponds roughly to the number of decimals of the predicted CDFs that are calculated accuratly.}

\item{fix_to}{List containing fixed parameter values. Names are parameters, elements are values.}

\item{depend_on_condition}{List containing parameters which may vary across conditions. Names are parameters, elements are conditions.}

\item{invariant}{Vector of names of parameters, which are not fixed but invariant across conditions.}

\item{delete}{Should temporary files be deleted?}
}
\value{
Returns a list of the input data, individual and aggregated parameter estimates as well as individual and aggregated cdf values per condition.
}
\description{
Estimate a diffusion model using fast-dm.
}
\examples{
# simulate data for 2 subjects
data = data.frame(sub = rep(c(1,2), each = 100),
                 cnd = rep(c(1,2), times = 100),
                 RESPONSE = sample (c(0,1), 200, p=c(0.1,0.9), replace = TRUE),
                 TIME = round((rnorm(200,400,30) + rexp(200,0.01))/1000, 2))

# load package
library(FastDMinR)

results <- fast_dm(data,
                  Subject = "sub",
                  Conditions = "cnd",
                  TIME = "TIME",
                  RESPONSE = "RESPONSE",
                  precision = 5.0,
                  method = "ks",
                  fix_to = list(p = 0, d = 0, sv = 0, st0 = 0, szr = 0),
                  depend_on_condition = list(a = "cnd"),
                  invariant = c("zr", "v", "t0"))

# plot cdf values with ggplot2
library(ggplot2)
ggplot(results$cdf$aggr_cdf, aes(x = RT, y = CDF)) +
 geom_line(aes(lty = cdf_Type), lwd = 1) +
 facet_grid(. ~ cnd)

}
\seealso{
For further information regarding fast-dm see \url{https://www.psychologie.uni-heidelberg.de/ae/meth/fast-dm/}.
}
\keyword{diffusion}
\keyword{fast-dm}
\keyword{modeling}
